athenaos-ai-service/requirements.txt
# requirements.txt

# Web framework
fastapi
uvicorn

# API Clients and Configuration
requests 
python-dotenv
groq